requirements.txt
torch==2.1.2
gradio==4.44.0
transformers==4.36.2
Pillow==10.2.0
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99
tensorflow==2.15.0
einops==0.7.0
numpy>=1.24.3
tqdm>=4.65.0
protobuf>=4.25.1
safetensors>=0.4.1
typing-extensions>=4.8.0
pyyaml>=6.0.1
requests>=2.31.0
packaging>=23.2
filelock>=3.13.1
regex>=2023.10.3
tokenizers>=0.15.0
huggingface-hub>=0.19.4
